REF: move anova_generic function to stats.oneway
